feat: add i18n store
