+++ apps/gitness/src/i18n/i18n.ts
@@ -6,14 +6,9 @@ import LanguageDetector from 'i18next-browser-languagedetector'
 import { resources } from '@harnessio/ui/locales'
 
 const languageDetectorOptions = {
-  // Order and from where user language should be detected
   order: ['localStorage', 'navigator', 'cookie'],
-
-  // Keys to search language in
   lookupCookie: 'i18next',
   lookupLocalStorage: 'i18nextLng',
-
-  // Cache user language on
   caches: ['cookie', 'localStorage']
 }
 
@@ -33,22 +28,6 @@ i18n
     }
   })
 
-export const handleLanguageChange = (lng: string) => {
-  // set system language
-  if (lng === 'system') {
-    const navigatorLang = (navigator.language || 'en').split('-')[0]
-    i18n.changeLanguage(navigatorLang)
-    localStorage.setItem('i18nextLngSystem', 'true')
-    localStorage.setItem('i18nextLng', navigatorLang)
-  } else {
-    // Set specific language
-    i18n.changeLanguage(lng)
-    localStorage.setItem('i18nextLng', lng)
-    localStorage.removeItem('i18nextLngSystem')
-  }
-}
-
-// Detect language change on the navigator
 window.addEventListener('languagechange', () => {
   const navigatorLang = (navigator.language || 'en').split('-')[0]
   i18n.changeLanguage(navigatorLang)
